requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
google-generativeai>=0.7.0
PyPDF2==3.0.1
pdfplumber==0.10.3
ebooklib==0.18
python-docx==1.1.0
Pillow==10.1.0
pytesseract==0.3.10
gTTS==2.4.0
SpeechRecognition==3.10.0
reportlab==4.0.7
wordcloud==1.9.3
Werkzeug==3.0.1
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
charset-normalizer==3.3.2
